chore: Removed previous grid components and replaced them with the new one
chore: Removed react-awesome-styled-grid typings

chore: Updated DropDown stories to use the new Grid

chore: Updated Snapshots for Grid
