+++ src/DropDown/DropDown.stories.tsx
@@ -11,7 +11,7 @@ import { DropDown, Trigger } from './DropDown.component';
 import { Button } from '../Button';
 import { List, Item as ListItem } from '../List';
 import { Typography } from '../Typography';
-import { Container, Row, Col } from '../Grid';
+import { Cell, Grid } from '../Grid';
 // THEME
 import { RootTheme } from '../theme';
 // README
@@ -56,129 +56,107 @@ storiesOf('Components/DropDown', module)
     .add('Default', () => (
         <ThemeProvider theme={RootTheme}>
             <StyledStory>
-                <Container>
-                    <Row>
-                        <Col offset={{ xs: 3, sm: 3, md: 3, lg: 3 }} lg={4}>
-                            <div>
-                                <Typography tag="h1">
-                                    DropDown Default
-                                </Typography>
-                                <DropDown overlay={<MockList />}>
+                <Grid gap="2rem">
+                    <Cell width={3}>
+                        <div>
+                            <Typography tag="h1">DropDown Default</Typography>
+                            <DropDown overlay={<MockList />}>
+                                <span>
                                     <span>
                                         <span>
-                                            <span>
-                                                {/* deeply nested item */}
-                                                <a>DropDown Link</a>
-                                            </span>
+                                            {/* deeply nested item */}
+                                            <a>DropDown Link</a>
                                         </span>
                                     </span>
-                                </DropDown>
-                            </div>
-                        </Col>
-                    </Row>
-                    <Row>
-                        <Col offset={{ xs: 3, sm: 3, md: 3, lg: 3 }} lg={4}>
-                            <div>
-                                <Typography tag="h1">DropDown Right</Typography>
-                                <DropDown
-                                    overlay={<MockList />}
-                                    position="right"
-                                >
-                                    <a>DropDown Link Right</a>
-                                </DropDown>
-                            </div>
-                        </Col>
-                    </Row>
-                    <Row>
-                        <Col offset={{ xs: 3, sm: 3, md: 3, lg: 3 }} lg={4}>
-                            <div>
-                                <Typography tag="h1">DropDown Left</Typography>
-                                <DropDown
-                                    overlay={<MockList />}
-                                    position="left"
-                                >
-                                    <a>DropDown Link Left</a>
-                                </DropDown>
-                            </div>
-                        </Col>
-                    </Row>
-                    <Row>
-                        <Col offset={{ xs: 3, sm: 3, md: 3, lg: 3 }} lg={4}>
-                            <div>
-                                <Typography tag="h1">DropDown Top</Typography>
-                                <DropDown
-                                    overlay={<MockList />}
-                                    position="top"
-                                    trigger="click"
-                                >
-                                    <a>DropDown Link Top</a>
-                                </DropDown>
-                            </div>
-                        </Col>
-                    </Row>
-                </Container>
+                                </span>
+                            </DropDown>
+                        </div>
+                    </Cell>
+                    <Cell width={3} top={2}>
+                        <div>
+                            <Typography tag="h1">DropDown Right</Typography>
+                            <DropDown overlay={<MockList />} position="right">
+                                <a>DropDown Link Right</a>
+                            </DropDown>
+                        </div>
+                    </Cell>
+                    <Cell width={3} top={3}>
+                        <div>
+                            <Typography tag="h1">DropDown Left</Typography>
+                            <DropDown overlay={<MockList />} position="left">
+                                <a>DropDown Link Left</a>
+                            </DropDown>
+                        </div>
+                    </Cell>
+                    <Cell width={3} top={4}>
+                        <div>
+                            <Typography tag="h1">DropDown Top</Typography>
+                            <DropDown
+                                overlay={<MockList />}
+                                position="top"
+                                trigger="click"
+                            >
+                                <a>DropDown Link Top</a>
+                            </DropDown>
+                        </div>
+                    </Cell>
+                </Grid>
             </StyledStory>
         </ThemeProvider>
     ))
     .add('Knobs Demo', () => (
         <ThemeProvider theme={RootTheme}>
             <StyledStory>
-                <Container>
-                    <Row>
-                        <Col offset={{ xs: 3, sm: 3, md: 3, lg: 3 }} lg={4}>
-                            <div>
-                                <Typography tag="h1">DropDown</Typography>
-                                <DropDown
-                                    overlay={<MockList />}
-                                    position={select<Position>(
-                                        'position',
-                                        [
-                                            'topStart',
-                                            'top',
-                                            'topEnd',
-                                            'rightStart',
-                                            'right',
-                                            'rightEnd',
-                                            'bottomEnd',
-                                            'bottom',
-                                            'bottomStart',
-                                            'leftEnd',
-                                            'left',
-                                            'leftStart',
-                                        ],
-                                        'bottom'
-                                    )}
-                                    trigger={select<Trigger>(
-                                        'trigger',
-                                        ['click', 'hover', 'both'],
-                                        'hover'
-                                    )}
-                                    showArrow={boolean('showArrow', true)}
-                                    shadow={text('shadow', '') || undefined}
-                                    background={
-                                        text('background', '') || undefined
-                                    }
-                                    border={text('border', '') || undefined}
-                                    borderRadius={
-                                        text('borderRadius', '') || undefined
-                                    }
-                                    arrowIndent={
-                                        text('arrowIndent', '') || undefined
-                                    }
-                                    arrowSize={
-                                        text('arrowSize', '') || undefined
-                                    }
-                                    spacing={text('spacing', '') || undefined}
-                                >
-                                    <Button>
-                                        {/* Not a prop, just for testing various widths */}
-                                        {text('Text', 'Dropdown')}
-                                    </Button>
-                                </DropDown>
-                            </div>
-                        </Col>
-                    </Row>
-                </Container>
+                <Grid>
+                    <Cell>
+                        <div>
+                            <Typography tag="h1">DropDown</Typography>
+                            <DropDown
+                                overlay={<MockList />}
+                                position={select<Position>(
+                                    'position',
+                                    [
+                                        'topStart',
+                                        'top',
+                                        'topEnd',
+                                        'rightStart',
+                                        'right',
+                                        'rightEnd',
+                                        'bottomEnd',
+                                        'bottom',
+                                        'bottomStart',
+                                        'leftEnd',
+                                        'left',
+                                        'leftStart',
+                                    ],
+                                    'bottom'
+                                )}
+                                trigger={select<Trigger>(
+                                    'trigger',
+                                    ['click', 'hover', 'both'],
+                                    'hover'
+                                )}
+                                showArrow={boolean('showArrow', true)}
+                                shadow={text('shadow', '') || undefined}
+                                background={text('background', '') || undefined}
+                                border={text('border', '') || undefined}
+                                borderRadius={
+                                    text('borderRadius', '') || undefined
+                                }
+                                arrowIndent={
+                                    text('arrowIndent', '') || undefined
+                                }
+                                arrowSize={text('arrowSize', '') || undefined}
+                                spacing={text('spacing', '') || undefined}
+                            >
+                                <Button>
+                                    {/* Not a prop, just for testing various widths */}
+                                    {text('Text', 'Dropdown')}
+                                </Button>
+                            </DropDown>
+                        </div>
+                    </Cell>
+                </Grid>
             </StyledStory>
         </ThemeProvider>
     ));
