--- src/Grid/Grid/README.md
@@ -1,1 +1,0 @@
-# Grid
