--- src/GridRefactor/Grid/index.ts
@@ -1,1 +1,0 @@
-export { Grid } from './Grid.component';
