--- src/GridRefactor/index.ts
@@ -1,2 +1,0 @@
-export { Cell } from './Cell';
-export { Grid } from './Grid';
