+++ src/typings.d.ts
@@ -133,45 +133,3 @@ declare module '@xstyled/system' {
     export const margin: StyleFn;
     export const padding: StyleFn;
 }
-
-declare module 'react-awesome-styled-grid' {
-    import { FunctionComponent, HTMLAttributes } from 'react';
-
-    type ColVal = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
-
-    interface ResponsiveProps {
-        xs?: ColVal;
-        sm?: ColVal;
-        md?: ColVal;
-        lg?: ColVal;
-        xl?: ColVal;
-    }
-    interface VisibleProps extends HTMLAttributes<HTMLDivElement> {
-        xs?: boolean;
-        sm?: boolean;
-        md?: boolean;
-        lg?: boolean;
-        xl?: boolean;
-        debug?: boolean;
-    }
-
-    type ResponsiveArrayValues = 'xs' | 'sm' | 'md' | 'lg' | 'xl';
-
-    interface ColProps extends ResponsiveProps, HTMLAttributes<HTMLDivElement> {
-        offset?: ResponsiveProps;
-        reverse?: ResponsiveArrayValues[];
-        debug?: boolean;
-        noGutter?: boolean;
-    }
-    interface ScreenClassProps extends HTMLAttributes<HTMLDivElement> {
-        render: (resolution: ResponsiveArrayValues) => any;
-    }
-
-    export const Row: FunctionComponent<HTMLAttributes<HTMLDivElement>>;
-    export const Col: FunctionComponent<ColProps>;
-    export const Container: FunctionComponent<HTMLAttributes<HTMLDivElement>>;
-    export const Visible: FunctionComponent<VisibleProps>;
-    export const Hidden: FunctionComponent<VisibleProps>;
-    export const ScreenClass: FunctionComponent<ScreenClassProps>;
-    export const ScreenBadge: FunctionComponent<HTMLAttributes<HTMLDivElement>>;
-}
